refactor: 상수 이동에 따른 코드 수정
